refactor(common): establish financial amount domain owner
