+++ src/libs/portfolio-common/portfolio_common/domain/financial/__init__.py
@@ -1,0 +1,1 @@
+"""Shared financial amount value objects and policies."""
